# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Error {
  error: string;
}

export interface ProblemCard {
  key: string;
  label: string;
  content: string;
}

export interface ProblemCardUpdate {
  content: string;
}

export interface Idea {
  id: number;
  text: string;
  author: string;
  votes: number;
  createdAt: string;
}

export interface IdeaInput {
  /** @minLength 1 */
  text: string;
  /** @minLength 1 */
  author: string;
}

export interface Task {
  id: number;
  weekNumber: number;
  name: string;
  owner: string;
  targetDate: string;
  done: boolean;
  position: number;
}

export interface TaskInput {
  /**
     * @minimum 1
     * @maximum 3
     */
  weekNumber: number;
  name?: string;
  owner?: string;
  targetDate?: string;
}

export interface TaskUpdate {
  name?: string;
  owner?: string;
  targetDate?: string;
  done?: boolean;
}

export interface Week {
  number: number;
  title: string;
  dateRange: string;
  theme: string;
  tasks: Task[];
  doneCount: number;
  totalCount: number;
}

export interface Plan {
  weeks: Week[];
  totalTasks: number;
  doneTasks: number;
  percentComplete: number;
}

export interface QcField {
  fieldName: string;
  fieldCaption: string;
  dataType?: string;
}

export interface QcConnectionStatus {
  connected: boolean;
  datasourceName?: string;
  fieldCount?: number;
  fields?: QcField[];
  error?: string;
}

export interface QcMetric {
  id: number;
  measureField: string;
  aggregation: string;
  dimensionField: string;
  thresholdPct: number;
  createdAt: string;
}

export interface QcMetricInput {
  /** @minLength 1 */
  measureField: string;
  aggregation?: string;
  dimensionField?: string;
  /** @minimum 0 */
  thresholdPct?: number;
}

export interface QcMetricUpdate {
  /** @minLength 1 */
  measureField?: string;
  aggregation?: string;
  dimensionField?: string;
  /** @minimum 0 */
  thresholdPct?: number;
}

export interface QcRun {
  id: number;
  trigger: string;
  status: string;
  message: string;
  metricsPulled: number;
  deviationsFound: number;
  startedAt: string;
}

export interface QcDeviation {
  id: number;
  runId: number;
  metricKey: string;
  measureField: string;
  dimensionValue: string;
  /** @nullable */
  previousValue?: number | null;
  /** @nullable */
  currentValue?: number | null;
  /** @nullable */
  pctChange?: number | null;
  reason: string;
  likelyCause: string;
  confidence: string;
  explanation: string;
  recommendedCheck: string;
  createdAt: string;
}

export interface QcRunResult {
  run: QcRun;
  deviations: QcDeviation[];
}

export type ListQcRunsParams = {
/**
 * @minimum 1
 * @maximum 100
 */
limit?: number;
};

export type ListQcDeviationsParams = {
/**
 * @minimum 1
 * @maximum 100
 */
limit?: number;
};