# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /qc/config:
    get:
      operationId: listQcMetrics
      tags: [qc]
      summary: List the monitored metric configurations
      responses:
        "200":
          description: List of monitored metrics
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/QcMetric"
    post:
      operationId: createQcMetric
      tags: [qc]
      summary: Add a monitored metric
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QcMetricInput"
      responses:
        "201":
          description: Created metric
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QcMetric"
  /qc/config/{id}:
    patch:
      operationId: updateQcMetric
      tags: [qc]
      summary: Update a monitored metric
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QcMetricUpdate"
      responses:
        "200":
          description: Updated metric
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QcMetric"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      operationId: deleteQcMetric
      tags: [qc]
      summary: Delete a monitored metric
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /qc/runs:
    get:
      operationId: listQcRuns
      tags: [qc]
      summary: List recent agent runs (most recent first)
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
      responses:
        "200":
          description: Recent runs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/QcRun"
  /qc/deviations:
    get:
      operationId: listQcDeviations
      tags: [qc]
      summary: List recent deviations (most recent first)
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
      responses:
        "200":
          description: Recent deviations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/QcDeviation"
...
components:
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    QcMetric:
      type: object
      required:
        [id, measureField, aggregation, dimensionField, thresholdPct, createdAt]
      properties:
        id:
          type: integer
        measureField:
          type: string
        aggregation:
          type: string
        dimensionField:
          type: string
        thresholdPct:
          type: number
        createdAt:
          type: string
          format: date-time
    QcMetricInput:
      type: object
      required: [measureField]
      properties:
        measureField:
          type: string
          minLength: 1
        aggregation:
          type: string
        dimensionField:
          type: string
        thresholdPct:
          type: number
          minimum: 0
    QcMetricUpdate:
      type: object
      properties:
        measureField:
          type: string
          minLength: 1
        aggregation:
          type: string
        dimensionField:
          type: string
        thresholdPct:
          type: number
          minimum: 0
    QcRun:
      type: object
      required:
        [id, trigger, status, message, metricsPulled, deviationsFound, startedAt]
      properties:
        id:
          type: integer
        trigger:
          type: string
        status:
          type: string
        message:
          type: string
        metricsPulled:
          type: integer
        deviationsFound:
          type: integer
        startedAt:
          type: string
          format: date-time
    QcDeviation:
      type: object
      required:
        [
          id,
          runId,
          metricKey,
          measureField,
          dimensionValue,
          reason,
          likelyCause,
          confidence,
          explanation,
          recommendedCheck,
          createdAt,
        ]
      properties:
        id:
          type: integer
        runId:
          type: integer
        metricKey:
          type: string
        measureField:
          type: string
        dimensionValue:
          type: string
        previousValue:
          type: [number, "null"]
        currentValue:
          type: [number, "null"]
        pctChange:
          type: [number, "null"]
        reason:
          type: string
        likelyCause:
          type: string
        confidence:
          type: string
        explanation:
          type: string
        recommendedCheck:
          type: string
        createdAt:
          type: string
          format: date-time
...