# lib/api-zod/src/generated/types/qcDeviation.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface QcDeviation {
  id: number;
  runId: number;
  metricKey: string;
  measureField: string;
  dimensionValue: string;
  /** @nullable */
  previousValue?: number | null;
  /** @nullable */
  currentValue?: number | null;
  /** @nullable */
  pctChange?: number | null;
  reason: string;
  likelyCause: string;
  confidence: string;
  explanation: string;
  recommendedCheck: string;
  createdAt: Date;
}