# lib/api-zod/src/generated/types/qcMetric.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface QcMetric {
  id: number;
  measureField: string;
  aggregation: string;
  dimensionField: string;
  thresholdPct: number;
  createdAt: Date;
}