# lib/api-zod/src/generated/types/qcMetricInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface QcMetricInput {
  /** @minLength 1 */
  measureField: string;
  aggregation?: string;
  dimensionField?: string;
  /** @minimum 0 */
  thresholdPct?: number;
}